+++ src/main/java/lotto/domain/ticket/TicketValidator.java
@@ -1,4 +1,4 @@
-package lotto.domain.ticket.validation;
+package lotto.domain.ticket;
 
 import java.util.Collections;
 import java.util.List;
