+++ src/main/java/lotto/utils/MoneyUnit.java
@@ -1,4 +1,4 @@
-package lotto.domain.money;
+package lotto.utils;
 
 public enum MoneyUnit {
